+++ metadata.rb
@@ -15,8 +15,10 @@ issues_url       'https://github.com/roboticcheese/dropbox-chef/issues'
 depends          'dmg', '~> 2.2'
 depends          'windows', '~> 1.36'
 depends          'yum', '~> 3.7'
+depends          'apt', '~> 2.8'
 
 supports         'mac_os_x'
 supports         'windows'
 supports         'fedora'
+supports         'ubuntu'
 # rubocop:enable SingleSpaceBeforeFirstArg