+++ test/integration/default/serverspec/localhost/package_spec.rb
@@ -15,8 +15,13 @@ describe 'Dropbox package' do
     end
   end
 
-  describe package('nautilus-dropbox'),
-           if: !%w(darwin windows).include?(os[:family]) do
+  describe package('dropbox'), if: os[:family] == 'ubuntu' do
+    it 'is installed' do
+      expect(subject).to be_installed
+    end
+  end
+
+  describe package('nautilus-dropbox'), if: os[:family] == 'fedora' do
     it 'is installed' do
       expect(subject).to be_installed
     end