+++ CHANGELOG.md
@@ -6,6 +6,8 @@ v?.?.? (????-??-??)
 - Rewrite to modern resource/provider standards (breaking compatibility with
   Chef < 12)
 - Change `package_url` attribute to `source`
+- Add support for Fedora Linux
+- Add support for Ubuntu Linux
 
 v0.1.0 (2014-12-27)
 -------------------